Organize functions and classes in files
